Toast support for prefab crate
