+++ .github/workflows/ci_develop.yaml
@@ -8,6 +8,10 @@ on:
     branches:
       - develop
 
+
+permissions:
+  contents: write
+  
 env:
   CARGO_TERM_COLOR: always
